+++ examples/ipv6/wsn-ping6.cc
@@ -88,7 +88,7 @@ main(int argc, char** argv)
     // Add and install the LrWpanNetDevice for each node
     // lrWpanHelper.EnableLogComponents();
     NetDeviceContainer devContainer = lrWpanHelper.Install(nodes);
-    lrWpanHelper.AssociateToPan(devContainer, 10);
+    lrWpanHelper.CreateAssociatedPan(devContainer, 10);
 
     std::cout << "Created " << devContainer.GetN() << " devices" << std::endl;
     std::cout << "There are " << nodes.GetN() << " nodes" << std::endl;